requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blessed==1.20.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
distlib==0.3.6
docker==7.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.12.2
fonttools==4.49.0
fsspec==2024.3.1
ftfy==6.1.3
google==3.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
gradio==4.22.0
gradio_client==0.13.0
grpcio==1.56.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.5
mdurl==0.1.2
murmurhash==1.0.10
notebook==5.2.2
numpy==1.24.4
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.9.0.80
orjson==3.9.15
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
portpicker==1.2.0
preshed==3.0.9
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
python-slugify==8.0.1
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.4
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminaltables==3.1.10
text-unidecode==1.3
thinc==8.2.2
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1+cu117
torch-tb-profiler==0.4.1
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.29.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
